--- src/widgets/mypage/index.ts
@@ -1,4 +1,0 @@
-export * from './likeProject';
-export * from './registeredProject';
-export * from './requestedProject';
-export type { GetMyResponse } from './model/types';
