--- src/widgets/mypage/likeProject/index.ts
@@ -1,1 +1,0 @@
-export { default as LikeProjects } from './ui/index';
