--- src/widgets/mypage/registeredProject/index.ts
@@ -1,1 +1,0 @@
-export { default as RegisteredProjects } from './ui/index';
