--- src/widgets/mypage/requestedProject/index.ts
@@ -1,1 +1,0 @@
-export { default as RequestedProjects } from './ui/index';
