--- src/widgets/mypage/requestedProject/model/index.ts
@@ -1,2 +1,0 @@
-export { default as useGetPending } from './useGetPending';
-export { default as useGetRejected } from './useGetRejected';
