+++ src/widgets/requestedProject/index.ts
@@ -1,0 +1,4 @@
+export { default as RequestedProjects } from './ui/index';
+export { default as useGetPending } from './model/useGetPending';
+export { default as useGetRejected } from './model/useGetRejected';
+export * from './model/types';
